Create ReadMe
So this is from a course offered on Udemy.  In this course, we used Class components with React to build a rolodex of monsters, and a search field.
The goal of the class was to teach the concepts of React.  The 'notes' file contains notes about hooks, and how they differ.  With the next commit,
I will follow the instructions to make this a hooks based project.